rename class and update tests
