--- src/aatoolbox/datasources/cds/__init__.py
@@ -1,1 +1,0 @@
-"""CDS API wrappers to download data from various sources."""
